Move performance history fresh start state management to performance-history
